+++ SimulationPr4/Pr.BL/Services/Concretes/DoctorService.cs
@@ -1,6 +1,7 @@
 ﻿using AutoMapper;
 using Microsoft.AspNetCore.Hosting;
 using Pr.BL.DTOs.DoctorDTOs;
+using Pr.BL.Exceptions;
 using Pr.BL.Services.Abstractions;
 using Pr.Core.Models;
 using Pr.DAL.Repositories.Abstractions;
@@ -33,6 +34,31 @@ namespace Pr.BL.Services.Concretes
         public async Task<Doctor> CreateAsync(DoctorCreateDTO entityDTo)
         {
             Doctor created = _mapper.Map<Doctor>(entityDTo);
+            string rootpath = _webHostEnvironment.WebRootPath;
+            string folder = rootpath + "/Uploads/Doctors";
+            if (!Directory.Exists(folder))
+            {
+                Directory.CreateDirectory(folder);
+            }
+            string fileName = entityDTo.Image.FileName;
+            string[] extensions = [".jpg", ".png", "jgeg"];
+            bool isAllowed = false;
+            foreach (var extension in extensions)
+            {
+                if (Path.GetExtension(fileName) == extension)
+                {
+                    isAllowed = true;
+                    break;
+                }
+
+            }
+            if (isAllowed) { throw new Exception("File is not sipported."); }
+            string filepath = folder + fileName;
+            using (FileStream stream = new FileStream(folder, FileMode.Create))
+            {
+                await entityDTo.Image.CopyToAsync(stream);
+            }
+            created.ImgURL = filepath;
             created.CreatedDate = DateTime.UtcNow.AddHours(4);
             await _repository.CreateAsync(created);
             await _repository.SaveChangesAsync();
@@ -42,9 +68,17 @@ namespace Pr.BL.Services.Concretes
         public async Task<Doctor> DeleteAsync(int id)
         {
            var entity = await _repository.GetByIdAsync(id);
-            entity.DeleteDate = DateTime.UtcNow.AddHours(4);
+            if (entity != null) 
+            {
+                throw new ItemNotFoundException("Item not found");
+            }
+            //entity.DeleteDate = DateTime.UtcNow.AddHours(4);
             await _repository.DeleteAsync(entity);
-            await _repository.SaveChangesAsync();
+           int result = await _repository.SaveChangesAsync();
+            if(result == 0)
+            {
+                throw new OperationCanceledException("Couldn't deleted doctor ");
+            }
             return entity;
         }
